requirements.txt
Counter
pycryptodome